Remove last use of .first in invoice generation
We want to group project resources and resource line items together and then
iterate over each array for cost calculation. Some fields in those arrays are
same since they are grouped together. We used to use .first when we need to
access those fields since any of them would work, however using .first is
confusing in this context as it implies .first element is special in some way.
With this commit, we eliminate the need of using .first. Instead we assign
the values we would need at the group_by phase.